Use last panels per side entry as default
There was a bug in the function to get the number of panels per side.
The total number of old panels was being passed to the function instead
of the old number of panels per side.

This commit remedies that bug and also uses the last number of panels
per side as the default value in the prompt, if there was one. If not,
it still uses the default value of 16.